fix: replace ts files with js files for cypress support
