+++ app/utils/formatHomePageEventDates.js
@@ -1,4 +1,4 @@
-import {format} from 'date-fns/format'
+import { format } from 'date-fns/format'
 
 /**
  * Format a single date for FTAV events
