+++ cypress/support/viewports.ts
@@ -2,4 +2,4 @@ export const viewports = [
   { label: 'Mobile', viewportWidth: 375, viewportHeight: 667 },
   { label: 'Tablet', viewportWidth: 768, viewportHeight: 1024 },
   { label: 'Desktop', viewportWidth: 1280, viewportHeight: 800 },
-] as const
+]
