+++ package.json
@@ -49,7 +49,6 @@
     "lodash": "^4.17.21",
     "nuxt-graphql-request": "8.1.1",
     "sass": "^1.66.1",
-    "typescript": "^7.0.2",
     "ucla-library-design-tokens": "6.0.2",
     "vite-svg-loader": "^5.1.0"
   },
